docs(decimal): add playground example
