+++ src/components/decimal/decimal.mdx
@@ -1,4 +1,4 @@
-import { Meta, ArgTypes, Canvas } from "@storybook/addon-docs/blocks";
+import { Meta, ArgTypes, Canvas, Controls } from "@storybook/addon-docs/blocks";
 import TranslationKeysTable from "../../../.storybook/utils/translation-keys-table";
 
 import * as DecimalStories from "./decimal.stories";
@@ -32,81 +32,41 @@ Captures a number with a decimal point, or a currency value.
 import Decimal from "carbon-react/lib/components/decimal";
 ```
 
+## Playground
+
+Use this interactive example to explore Decimal props with Storybook controls.
+
+<Canvas of={DecimalStories.Playground} />
+
+<Controls
+  of={DecimalStories.Playground}
+  include={[
+    "label",
+    "required",
+    "disabled",
+    "readOnly",
+    "size",
+    "prefix",
+    "suffix",
+    "precision",
+    "inputHint",
+    "allowEmptyValue",
+    "align",
+    "locale",
+    "inputWidth",
+    "maxWidth",
+    "labelInline",
+    "error",
+  ]}
+/>
+
 ## Examples
-
-### Default
-
-<Canvas of={DecimalStories.DefaultStory} />
-
-### With Input Hint
-
-<Canvas of={DecimalStories.WithInputHint} />
-
-### Sizes
-
-<Canvas of={DecimalStories.Sizes} />
-
-### Disabled
-
-<Canvas of={DecimalStories.Disabled} />
-
-### Prefix
-
-<Canvas of={DecimalStories.Prefix} />
-
-### Suffix
-
-<Canvas of={DecimalStories.Suffix} />
-
-**Note:** `prefix` takes precedence over `suffix`.
 
 ### With Popover Container
 
 Use the `popoverContainerContent` prop to render content inside a [PopoverContainer](../?path=/docs/popover-container--docs) accessible via a button at the right of the input. Use the `popoverPosition` prop to control the position of the popover dialog relative to the trigger button.
 
-<Canvas of={DecimalStories.WithPopoverContainer} /> If both props are supplied, only the prefix will be rendered and a console warning will be logged. Use one or the other.
-
-### ReadOnly
-
-<Canvas of={DecimalStories.ReadOnly} />
-
-### Empty
-
-<Canvas of={DecimalStories.Empty} />
-
-### With Custom Precision
-
-<Canvas of={DecimalStories.WithCustomPrecision} />
-
-### With custom maxWidth
-
-<Canvas of={DecimalStories.WithCustomMaxWidth} />
-
-### Required
-
-You can use the `required` prop to indicate if the field is mandatory.
-
-<Canvas of={DecimalStories.Required} />
-
-### Left aligned
-
-You can use the `align` prop to choose how the the characters inside the component align.
-In this example, `align` has been assigned the value `left`.
-
-<Canvas of={DecimalStories.LeftAligned} />
-
-### With labelInline
-
-Use the `labelInline` prop to display the label on the same horizontal row as the input. You can adjust its appearance using the `labelWidth` and 
-`labelAlign` props to control the width and text alignment of the label and the `inputWidth` prop to control the width of the input.
-
-<Canvas of={DecimalStories.LabelInline} />
-
-### With fieldHelp (legacy)
-
-**Note:** This is a legacy feature and will only render if the `validationRedesignOptIn` feature flag on an ancestor [CarbonProvider](../?path=/docs/carbon-provider--docs) is *false*.
-
-<Canvas of={DecimalStories.WithFieldHelp} />
+<Canvas of={DecimalStories.WithPopoverContainer} />
 
 ## Validation States
